apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.0
python-dotenv==1.0.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
pypdf==4.3.1
python-docx==1.1.2
requests==2.32.3

